+++ packages/sui-ssr/server/criticalCss/index.js
@@ -4,9 +4,8 @@ import {match} from 'react-router'
 import https from 'https'
 import parser from 'ua-parser-js'
 
-const PRODUCTION = 'production'
-const {NODE_ENV = PRODUCTION} = process.env
-const __CACHE__ = {}
+let __REQUESTING__ = false
+let __CACHE__ = {}
 
 const generateMinimalCSSHash = routes => {
   return routes.reduce((acc, route) => {
@@ -15,11 +14,22 @@ const generateMinimalCSSHash = routes => {
   }, '')
 }
 
-const logMessage = message => NODE_ENV !== PRODUCTION && console.log(message)
+const logMessageFactory = url => message =>
+  process.env.VERBOSE &&
+  console.log(`\u001b[36m[CRITICAL CSS](${url})\u001b[0m`, message)
 
 export default config => (req, res, next) => {
-  if (!config || process.env.DISABLE_CRITICAL_CSS === 'true') {
+  const logMessage = logMessageFactory(req.url)
+
+  if (req.skipSSR || !config || process.env.DISABLE_CRITICAL_CSS === 'true') {
+    logMessage('Skip middleware because it is inactive')
     return next()
+  }
+
+  if (req.url.match('x-criticalcss-cache-invalidate')) {
+    __CACHE__ = {}
+
+    logMessage('CriticalCSS cache invalidated')
   }
 
   if (
@@ -28,6 +38,7 @@ export default config => (req, res, next) => {
     Array.isArray(config.blackListURLs) &&
     config.blackListURLs.some(regex => req.url.match(regex))
   ) {
+    logMessage('Skip middleware because url is blacklisted')
     return next()
   }
 
@@ -60,7 +71,7 @@ export default config => (req, res, next) => {
       const hash = generateMinimalCSSHash(renderProps.routes) + '|' + device
       const criticalCSS = __CACHE__[hash]
 
-      if (!criticalCSS) {
+      if (!criticalCSS && !__REQUESTING__) {
         logMessage(`Generation Critical CSS for -> ${urlRequest} with ${hash}`)
 
         const serviceRequestURL = `https://critical-css-service.now.sh/${device}/${urlRequest}`
@@ -69,9 +80,13 @@ export default config => (req, res, next) => {
           ...(headers && {headers})
         }
 
+        logMessage(serviceRequestURL)
+
+        __REQUESTING__ = true
         https.get(serviceRequestURL, options, res => {
           let css = ''
           if (res.statusCode !== 200) {
+            __REQUESTING__ = false
             logMessage(`No 200 request, statusCode: ${res.statusCode}`)
 
             return
@@ -79,8 +94,15 @@ export default config => (req, res, next) => {
           res.on('data', data => {
             css += data
           })
+
+          res.on('error', () => {
+            logMessage(`Error Requesting ${serviceRequestURL}`)
+            __REQUESTING__ = false
+          })
+
           res.on('end', () => {
             logMessage(`Add cache entry for ${hash}`)
+            __REQUESTING__ = false
             __CACHE__[hash] = css
           })
         })